MAde better file structure
