+++ src/components/Card/ItemTypes.js
@@ -1,0 +1,3 @@
+export const ItemTypes = {
+  CARD: 'card',
+};
